+++ src/eca/config.clj
@@ -46,6 +46,7 @@
                          :requiresAuth? true
                          :models {"gpt-5.2" {}
                                   "gpt-5.2-codex" {}
+                                  "gpt-5.3-codex" {}
                                   "gpt-5-mini" {}
                                   "gpt-5-nano" {}
                                   "gpt-4.1" {}}}